# api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "校园活动管理系统 API",
        "description": "校园活动管理系统接口文档",
        "contact": {
            "name": "开发团队",
            "email": "example@xidian.edu.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
...
    "paths": {
...
        "/activity-types/{typeId}": {
            "get": {
                "tags": [
                    "活动类型管理"
                ],
                "summary": "获取活动类型详情",
                "description": "根据ID获取活动类型详细信息",
                "operationId": "getActivityTypeDetail",
                "parameters": [
                    {
                        "name": "typeId",
                        "in": "path",
                        "description": "活动类型ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityTypeVO"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "活动类型管理"
                ],
                "summary": "修改活动类型",
                "description": "修改活动类型信息",
                "operationId": "updateActivityType",
                "parameters": [
                    {
                        "name": "typeId",
                        "in": "path",
                        "description": "活动类型ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityTypeUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityTypeVO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "活动类型管理"
                ],
                "summary": "删除活动类型",
                "description": "删除活动类型（软删除）",
                "operationId": "deleteActivityType",
                "parameters": [
                    {
                        "name": "typeId",
                        "in": "path",
                        "description": "活动类型ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/activity/create": {
            "post": {
                "tags": [
                    "活动管理"
                ],
                "summary": "创建活动",
                "description": "管理员创建新的活动",
                "operationId": "createActivity",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityCreateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityDetailVO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/activity-types/create": {
            "post": {
                "tags": [
                    "活动类型管理"
                ],
                "summary": "创建活动类型",
                "description": "管理员创建新的活动类型",
                "operationId": "createActivityType",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivityTypeCreateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityTypeVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activity-types/batch-delete": {
            "post": {
                "tags": [
                    "活动类型管理"
                ],
                "summary": "批量删除活动类型",
                "description": "批量删除活动类型",
                "operationId": "batchDeleteActivityTypes",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "array",
                                "description": "活动类型ID列表",
                                "items": {
                                    "type": "integer",
                                    "format": "int64"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultInteger"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/activity/{activityId}": {
            "get": {
                "tags": [
                    "活动管理"
                ],
                "summary": "获取活动详情",
                "description": "根据活动ID获取详细信息",
                "operationId": "getActivityDetail",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "活动ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityDetailVO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "活动管理"
                ],
                "summary": "删除活动",
                "description": "删除指定活动",
                "operationId": "deleteActivity",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "活动ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activity/list": {
            "get": {
                "tags": [
                    "活动管理"
                ],
                "summary": "查询活动列表",
                "description": "分页查询活动列表，支持筛选和排序",
                "operationId": "getActivityList",
                "parameters": [
                    {
                        "name": "queryDTO",
                        "in": "query",
                        "description": "查询条件",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/ActivityQueryDTO"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultPageInfoActivityListVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activity-types/list": {
            "get": {
                "tags": [
                    "活动类型管理"
                ],
                "summary": "查询活动类型列表",
                "description": "分页查询活动类型列表，支持筛选",
                "operationId": "getActivityTypeList",
                "parameters": [
                    {
                        "name": "queryDTO",
                        "in": "query",
                        "description": "查询条件",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/ActivityTypeQueryDTO"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultPageInfoActivityTypeVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activity-types/check/type-name": {
            "get": {
                "tags": [
                    "活动类型管理"
                ],
                "summary": "检查类型名称是否存在",
                "description": "检查类型名称是否已存在",
                "operationId": "checkTypeNameExists",
                "parameters": [
                    {
                        "name": "typeName",
                        "in": "query",
                        "description": "类型名称",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "excludeId",
                        "in": "query",
                        "description": "排除的类型ID（用于修改时检查）",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultBoolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activity-types/check/type-code": {
            "get": {
                "tags": [
                    "活动类型管理"
                ],
                "summary": "检查类型编码是否存在",
                "description": "检查类型编码是否已存在",
                "operationId": "checkTypeCodeExists",
                "parameters": [
                    {
                        "name": "typeCode",
                        "in": "query",
                        "description": "类型编码",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "excludeId",
                        "in": "query",
                        "description": "排除的类型ID（用于修改时检查）",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultBoolean"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ResultVoid": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
...
            "ResultActivityDetailVO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ActivityDetailVO"
                    }
                }
            },
...
            "ActivityTypeUpdateDTO": {
                "required": [
                    "typeCode",
                    "typeName"
                ],
                "type": "object",
                "properties": {
                    "typeName": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string",
                        "description": "类型名称",
                        "example": "学术讲座"
                    },
                    "typeCode": {
                        "pattern": "^[A-Z_]{2,20}$",
                        "type": "string",
                        "description": "类型编码",
                        "example": "ACADEMIC_LECTURE"
                    },
                    "sortOrder": {
                        "minimum": 0,
                        "type": "integer",
                        "description": "排序顺序，数字越小越靠前",
                        "format": "int32",
                        "example": 1
                    },
                    "isEnabled": {
                        "type": "integer",
                        "description": "是否启用，1为启用，0为禁用",
                        "format": "int32",
                        "example": 1
                    }
                },
                "description": "活动类型信息"
            },
...
            "ResultActivityTypeVO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ActivityTypeVO"
                    }
                }
            },
...
            "ActivityTypeCreateDTO": {
                "required": [
                    "isEnabled",
                    "sortOrder",
                    "typeCode",
                    "typeName"
                ],
                "type": "object",
                "properties": {
                    "typeName": {
                        "maxLength": 50,
                        "minLength": 0,
                        "type": "string",
                        "description": "类型名称",
                        "example": "学术讲座"
                    },
                    "typeCode": {
                        "pattern": "^[A-Z_]{2,20}$",
                        "type": "string",
                        "description": "类型编码",
                        "example": "ACADEMIC_LECTURE"
                    },
                    "sortOrder": {
                        "minimum": 0,
                        "type": "integer",
                        "description": "排序顺序，数字越小越靠前",
                        "format": "int32",
                        "example": 1
                    },
                    "isEnabled": {
                        "type": "integer",
                        "description": "是否启用，1为启用，0为禁用",
                        "format": "int32",
                        "example": 1
                    }
                },
                "description": "活动类型信息"
            },
            "ResultInteger": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "ResultPageInfoActivityListVO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageInfoActivityListVO"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer Authentication": {
                "type": "http",
                "description": "请输入JWT Token，格式：直接输入token即可（不需要Bearer前缀）",
                "name": "Bearer Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}